// src/app/model/core-domain-value.ts
export interface CoreDomainValue {
  id?: string;
  code: string;
  description: string;
  val1?: string;
  val2?: string;
  val3?: string;
  val4?: string;
  val5?: string;
  val6?: string;
  val7?: string;
  val8?: string;
  val9?: string;
  val10?: string;
  val11?: string;
  coreDomainId?: string;
  sysCreatedDate?: Date;
  sysCreatedBy?: string;
  sysUpdatedDate?: Date;
  sysUpdatedBy?: string;
  sysActiveFlag?: boolean;
}